+++ app/crud/__init__.py
@@ -1,3 +1,4 @@
+from .crud_drs_yield import drs_yield
 from .crud_farm import farm
 from .crud_field import field
 from .crud_research import research