+++ app/db/init_db.py
@@ -1,4 +1,5 @@
 import argparse
+import datetime
 import json
 import logging
 
@@ -19,6 +20,30 @@ settings = get_settings()
 parser = argparse.ArgumentParser()
 parser.add_argument("--testing", action="store_true")
 args = parser.parse_args()
+
+
+def parse_date(date_string):
+    """
+    Converts a date string in 'MM/DD/YY' format to a datetime.date object.
+
+    Args:
+    date_string (str): The date string to parse, expected in 'MM/DD/YY' format.
+
+    Returns:
+    datetime.date: A date object representing the parsed date.
+
+    Raises:
+    ValueError: If the date_string is not in the expected format or is invalid.
+    """
+    # Parse the date string using strptime with the appropriate format
+    try:
+        date_object = datetime.datetime.strptime(date_string, "%m/%d/%y").date()
+        return date_object
+    except ValueError as e:
+        # Raise an error with a more descriptive message
+        raise ValueError(
+            f"Invalid date or format: {date_string}. Expected format is MM/DD/YY."
+        ) from e
 
 
 class Data:
@@ -98,7 +123,23 @@ class Data:
                     "field_ref_id": field_in_db.id,
                 }
 
-                crud.research.create(self.db, obj_in=obj_in)
+                research = crud.research.create(self.db, obj_in=obj_in)
+
+                # Insert Drought Resistant Seed Data if available
+                if len(field["drs_yield_data"]) > 0:
+                    logger.info(
+                        f"Importing DRS Yield Data for research: {field['researchName']}"
+                    )
+                    for drs_yield in field["drs_yield_data"]:
+                        obj_in = {
+                            "replicate": drs_yield["replicate"],
+                            "line": drs_yield["line"],
+                            "planting_date": parse_date(drs_yield["planting_date"]),
+                            "harvest_date": parse_date(drs_yield["harvest_date"]),
+                            "crop_yield": drs_yield["crop_yield"],
+                            "research_ref_id": research.id,
+                        }
+                        crud.drs_yield.create(self.db, obj_in=obj_in)
 
                 # Insert Sensors
                 for sensor in field["sensors"]: