+++ app/models/__init__.py
@@ -1,3 +1,4 @@
+from .drs_yield import DroughtResistantSeedYield
 from .farm import Farm
 from .field import Field
 from .research import Research